fix: resolve issue #666 - refactor NewCauseClient into smaller testable units
- Extract validation logic into src/lib/campaignValidation.ts
- Extract draft management into src/hooks/useCampaignFormDraft.ts hook
- Extract review modal into src/components/CampaignReviewModal.tsx
- Reduce NewCauseClient.tsx from 1,187 to 895 lines
- Fix draft overwrite bug on initial mount with hasRestored guard
- Add explicit CampaignFormSetters type
- Fix pre-existing type errors in markdownSanitizeSchema.ts
