+++ src/app/[locale]/causes/new/NewCauseClient.tsx
@@ -2,7 +2,8 @@
 
 import { useTranslations } from "next-intl";
 import SafeMarkdown from "@/components/SafeMarkdown";
-import { useState, useEffect } from "react";
+import CampaignReviewModal from "@/components/CampaignReviewModal";
+import { useState, useEffect, useCallback } from "react";
 import { useToast } from "@/components/ToastProvider";
 import { useWallet } from "@/components/WalletContext";
 import { useRouter } from "@/i18n/routing";
@@ -15,93 +16,8 @@ import { Category, CATEGORY_LABELS } from "@/types";
 import { xlmToStroops } from "@/lib/stellarAmount";
 import { parseContractError } from "@/utils/contractErrors";
 import { encodeLocalizedDescription } from "@/utils/localizedDescription";
-
-// ---------------------------------------------------------------------------
-// Validation — returns translation keys instead of hardcoded strings
-// ---------------------------------------------------------------------------
-
-interface FormErrorKeys {
-  title?: string;
-  description?: string;
-  descriptionEs?: string;
-  creatorEmail?: string;
-  fundingGoal?: string;
-  durationDays?: string;
-  revenueSharePercentage?: string;
-  coverImageUrl?: string;
-}
-
-interface ReviewData {
-  title: string;
-  description: string;
-  creatorEmail: string;
-  fundingGoalXlm: number;
-  durationDays: number;
-  category: Category;
-  hasRevenueSharing: boolean;
-  revenueSharePercentage: number;
-  estimatedDeadlineTimestamp: number;
-  tags: string[];
-  coverImageUrl: string;
-  milestones: { targetAmount: bigint; description: string }[];
-}
-
-const IMAGE_URL_RE = /^https?:\/\/.+\..+/;
-
-function validateForm(
-  title: string,
-  description: string,
-  descriptionEs: string,
-  creatorEmail: string,
-  fundingGoal: string,
-  durationDays: string,
-  hasRevenueSharing: boolean,
-  revenueSharePercentage: number,
-  tags: string[],
-  coverImageUrl: string,
-): FormErrorKeys {
-  const errors: FormErrorKeys = {};
-
-  if (title.trim().length < 1) {
-    errors.title = "validationTitleRequired";
-  } else if (title.trim().length > 100) {
-    errors.title = "validationTitleTooLong";
-  }
-
-  if (description.trim().length < 1) {
-    errors.description = "validationDescriptionRequired";
-  } else if (description.trim().length > 1000) {
-    errors.description = "validationDescriptionTooLong";
-  }
-
-  if (descriptionEs.trim().length > 1000) {
-    errors.descriptionEs = "validationDescriptionEsTooLong";
-  }
-
-  if (creatorEmail.trim() && !/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(creatorEmail.trim())) {
-    errors.creatorEmail = "validationCreatorEmailInvalid";
-  }
-
-  const goal = parseFloat(fundingGoal);
-  if (!fundingGoal || isNaN(goal) || goal <= 0) {
-    errors.fundingGoal = "validationFundingGoalInvalid";
-  }
-
-  const days = parseInt(durationDays, 10);
-  if (!durationDays || isNaN(days) || days < 1 || days > 365) {
-    errors.durationDays = "validationDurationInvalid";
-  }
-
-  if (hasRevenueSharing && (revenueSharePercentage < 0.01 || revenueSharePercentage > 50)) {
-    errors.revenueSharePercentage = "validationRevenueShareInvalid";
-  }
-
-  if (coverImageUrl.trim() && !IMAGE_URL_RE.test(coverImageUrl.trim())) {
-    errors.coverImageUrl = "validationCoverImageInvalid";
-  }
-
-  return errors;
-}
+import { useCampaignFormDraft, type CampaignFormDraftData } from "@/hooks/useCampaignFormDraft";
+import { validateForm, type FormErrorKeys, type ReviewData } from "@/lib/campaignValidation";
 
 // ---------------------------------------------------------------------------
 // Component
@@ -113,6 +29,7 @@ export default function CreateCampaignPage() {
   const { publicKey, isWalletConnected, connectWallet, isLoading: walletLoading } = useWallet();
   const { showError, showSuccess, showWarning } = useToast();
 
+  // ── Form state ──
   const [title, setTitle] = useState("");
   const [description, setDescription] = useState("");
   const [creatorEmail, setCreatorEmail] = useState("");
@@ -120,7 +37,6 @@ export default function CreateCampaignPage() {
   const [durationDays, setDurationDays] = useState("");
   const [category, setCategory] = useState<Category>(Category.Learner);
   const [hasRevenueSharing, setHasRevenueSharing] = useState(false);
-  // #110 — default 5 % (500 bps); state is percent, converted to bps at submit
   const [revenueSharePercentage, setRevenueSharePercentage] = useState(5);
   const [errorKeys, setErrorKeys] = useState<FormErrorKeys>({});
   const [isSubmitting, setIsSubmitting] = useState(false);
@@ -134,98 +50,62 @@ export default function CreateCampaignPage() {
   const [coverImageUrl, setCoverImageUrl] = useState("");
   const [txPhase, setTxPhase] = useState<TransactionLifecyclePhase | null>(null);
 
-  const draftKey = publicKey
-    ? `proof_of_heart_draft_${publicKey}`
-    : "proof_of_heart_draft_anonymous";
-  const [hasDraft, setHasDraft] = useState(false);
-  const [lastSavedAt, setLastSavedAt] = useState<number | null>(null);
   const CREATOR_EMAIL_WEBHOOK_URL = process.env.NEXT_PUBLIC_CREATOR_EMAIL_WEBHOOK_URL?.trim() ?? "";
 
+  // ── Draft management hook ──
+  const { hasDraft, lastSavedAt, restoreDraft, saveDraft, discardDraft, clearDraft } =
+    useCampaignFormDraft({ publicKey });
+
+  // Restore draft on mount
   useEffect(() => {
-    try {
-      const saved = localStorage.getItem(draftKey);
-      if (saved) {
-        const parsed = JSON.parse(saved);
-        if (parsed.title) setTitle(parsed.title);
-        if (parsed.description) setDescription(parsed.description);
-        if (parsed.descriptionEs) setDescriptionEs(parsed.descriptionEs);
-        if (parsed.creatorEmail) setCreatorEmail(parsed.creatorEmail);
-        if (parsed.fundingGoal) setFundingGoal(parsed.fundingGoal);
-        if (parsed.durationDays) setDurationDays(parsed.durationDays);
-        if (parsed.category !== undefined) setCategory(parsed.category);
-        if (parsed.hasRevenueSharing !== undefined) setHasRevenueSharing(parsed.hasRevenueSharing);
-        if (parsed.revenueSharePercentage !== undefined)
-          setRevenueSharePercentage(parsed.revenueSharePercentage);
-        if (parsed.tags) setTags(parsed.tags);
-        if (parsed.coverImageUrl) setCoverImageUrl(parsed.coverImageUrl);
-        if (parsed.milestones) setMilestones(parsed.milestones);
-        setHasDraft(true);
-      }
-    } catch (e) {
-      console.warn("Failed to load draft from localStorage:", e);
-    }
+    restoreDraft({
+      setTitle,
+      setDescription,
+      setDescriptionEs,
+      setCreatorEmail,
+      setFundingGoal,
+      setDurationDays,
+      setCategory,
+      setHasRevenueSharing,
+      setRevenueSharePercentage,
+      setTags,
+      setCoverImageUrl,
+      setMilestones,
+    });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
+  // Persist draft whenever form state changes
   useEffect(() => {
-    try {
-      localStorage.setItem(
-        draftKey,
-        JSON.stringify({
-          title,
-          description,
-          descriptionEs,
-          creatorEmail,
-          fundingGoal,
-          durationDays,
-          category,
-          hasRevenueSharing,
-          revenueSharePercentage,
-          tags,
-          coverImageUrl,
-          milestones,
-        }),
-      );
-      setLastSavedAt(Date.now());
-      setHasDraft(true);
-    } catch (e) {
-      console.warn("Failed to save draft to localStorage:", e);
-    }
+    saveDraft({
+      title,
+      description,
+      descriptionEs,
+      creatorEmail,
+      fundingGoal,
+      durationDays,
+      category,
+      hasRevenueSharing,
+      revenueSharePercentage,
+      tags,
+      coverImageUrl,
+      milestones,
+    });
   }, [
-    title,
-    description,
-    descriptionEs,
-    creatorEmail,
-    fundingGoal,
-    durationDays,
-    category,
-    hasRevenueSharing,
-    revenueSharePercentage,
-    tags,
-    coverImageUrl,
-    milestones,
+    title, description, descriptionEs, creatorEmail, fundingGoal, durationDays,
+    category, hasRevenueSharing, revenueSharePercentage, tags, coverImageUrl, milestones,
+    saveDraft,
   ]);
 
   const handleDiscardDraft = () => {
-    try {
-      localStorage.removeItem(draftKey);
-    } catch {
-      // ignore
-    }
-    setTitle("");
-    setDescription("");
-    setCreatorEmail("");
-    setFundingGoal("");
-    setDurationDays("");
-    setCategory(Category.Learner);
-    setHasRevenueSharing(false);
-    setRevenueSharePercentage(5);
-    setTags([]);
-    setCoverImageUrl("");
-    setMilestones([]);
-    setHasDraft(false);
-    setLastSavedAt(null);
+    discardDraft({
+      setTitle, setDescription, setDescriptionEs, setCreatorEmail,
+      setFundingGoal, setDurationDays, setCategory, setHasRevenueSharing,
+      setRevenueSharePercentage, setTags, setCoverImageUrl, setMilestones,
+    });
   };
 
+  // ── Derived state ──
   const isStartup = category === Category.EducationalStartup;
 
   const handleCategoryChange = (val: Category) => {
@@ -236,6 +116,7 @@ export default function CreateCampaignPage() {
     }
   };
 
+  // ── Helpers ──
   const formatReviewDate = (timestamp: number) =>
     new Intl.DateTimeFormat(undefined, {
       year: "numeric",
@@ -246,36 +127,38 @@ export default function CreateCampaignPage() {
       timeZoneName: "short",
     }).format(new Date(timestamp * 1000));
 
-  const notifyEmailOptIn = async (
-    campaignId: number | null,
-    email: string,
-    campaignTitle: string,
-    creatorAddress: string,
-  ) => {
-    if (!CREATOR_EMAIL_WEBHOOK_URL || !email) return;
+  const notifyEmailOptIn = useCallback(
+    async (
+      campaignId: number | null,
+      email: string,
+      campaignTitle: string,
+      creatorAddress: string,
+    ) => {
+      if (!CREATOR_EMAIL_WEBHOOK_URL || !email) return;
+      try {
+        await fetch(CREATOR_EMAIL_WEBHOOK_URL, {
+          method: "POST",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({
+            event: "campaign_creator_email_opt_in",
+            email,
+            campaignId,
+            campaignTitle,
+            creatorAddress,
+            source: "proof_of_heart_frontend",
+            timestamp: new Date().toISOString(),
+          }),
+        });
+      } catch {
+        showWarning(t("emailWebhookFailed"));
+      }
+    },
+    [CREATOR_EMAIL_WEBHOOK_URL, showWarning, t],
+  );
 
-    try {
-      await fetch(CREATOR_EMAIL_WEBHOOK_URL, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          event: "campaign_creator_email_opt_in",
-          email,
-          campaignId,
-          campaignTitle,
-          creatorAddress,
-          source: "proof_of_heart_frontend",
-          timestamp: new Date().toISOString(),
-        }),
-      });
-    } catch {
-      showWarning(t("emailWebhookFailed"));
-    }
-  };
-
-  const handleConfirmAndSign = async () => {
+  // ── Submission handlers ──
+  const handleConfirmAndSign = useCallback(async () => {
     if (!reviewData) return;
-
     if (!isWalletConnected || !publicKey) {
       showError(t("walletRequiredError"));
       return;
@@ -310,7 +193,7 @@ export default function CreateCampaignPage() {
       try {
         newCampaignId = await getCampaignCount();
       } catch {
-        // Ignore count lookup failures and continue with a generic redirect.
+        // Ignore count lookup failures
       }
 
       await notifyEmailOptIn(newCampaignId, reviewData.creatorEmail, reviewData.title, publicKey);
@@ -318,27 +201,23 @@ export default function CreateCampaignPage() {
       showSuccess(t("successMessage", { title: reviewData.title }));
       setIsReviewOpen(false);
       setReviewData(null);
-
-      try {
-        localStorage.removeItem(draftKey);
-      } catch {
-        // ignore
-      }
+      clearDraft();
 
       if (newCampaignId !== null) {
         router.push(`/causes/${newCampaignId}`);
       } else {
         router.push("/causes");
       }
-      // Keep isSubmitting true so the form stays disabled during navigation.
-      // The component will unmount before this matters on the success path.
       return;
     } catch (err) {
       showError(parseContractError(err));
     }
     setIsSubmitting(false);
     setTxPhase(null);
-  };
+  }, [
+    reviewData, isWalletConnected, publicKey, showError, showSuccess, t,
+    notifyEmailOptIn, clearDraft, router,
+  ]);
 
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
@@ -349,16 +228,8 @@ export default function CreateCampaignPage() {
     }
 
     const keys = validateForm(
-      title,
-      description,
-      descriptionEs,
-      creatorEmail,
-      fundingGoal,
-      durationDays,
-      hasRevenueSharing,
-      revenueSharePercentage,
-      tags,
-      coverImageUrl,
+      title, description, descriptionEs, creatorEmail, fundingGoal,
+      durationDays, hasRevenueSharing, revenueSharePercentage, coverImageUrl,
     );
 
     if (Object.keys(keys).length > 0) {
@@ -398,14 +269,11 @@ export default function CreateCampaignPage() {
     setIsReviewOpen(true);
   };
 
-  // Resolve a key to a translated string (or undefined)
+  // Resolve error key to translated string
   const err = (key: keyof FormErrorKeys) =>
     errorKeys[key] ? t(errorKeys[key] as Parameters<typeof t>[0]) : undefined;
 
-  // ---------------------------------------------------------------------------
-  // Render
-  // ---------------------------------------------------------------------------
-
+  // ── Render ──
   return (
     <div className="min-h-screen bg-linear-to-br from-zinc-50 to-zinc-100 dark:from-zinc-900 dark:to-zinc-800">
       <main className="container mx-auto px-4 py-10 max-w-2xl">
@@ -599,7 +467,7 @@ export default function CreateCampaignPage() {
             </div>
           </div>
 
-          {/* Optional creator email (off-chain only) */}
+          {/* Optional creator email */}
           <div>
             <label
               htmlFor="creatorEmail"
@@ -636,7 +504,6 @@ export default function CreateCampaignPage() {
 
           {/* Funding Goal + Duration */}
           <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-            {/* Funding Goal */}
             <div>
               <label
                 htmlFor="fundingGoal"
@@ -672,7 +539,6 @@ export default function CreateCampaignPage() {
               )}
             </div>
 
-            {/* Duration */}
             <div>
               <label
                 htmlFor="durationDays"
@@ -763,10 +629,6 @@ export default function CreateCampaignPage() {
               </div>
 
               {hasRevenueSharing && (
-                // #110 — slider now models percent directly (0.5 % steps) so
-                // round percentages are easy to hit. A companion bps input lets
-                // users enter exact values when they need them. Submitted value
-                // is still converted to bps at submit time.
                 <div>
                   <div className="flex items-center justify-between mb-2">
                     <label
@@ -783,7 +645,6 @@ export default function CreateCampaignPage() {
                     </span>
                   </div>
 
-                  {/* Slider — operates in percent, 0.5 % steps (#110) */}
                   <input
                     id="revenueShareSlider"
                     type="range"
@@ -800,7 +661,6 @@ export default function CreateCampaignPage() {
                     <span>50%</span>
                   </div>
 
-                  {/* Precise bps input for exact values (#110) */}
                   <div className="flex items-center gap-2 mt-2">
                     <label
                       htmlFor="revenueShareBps"
@@ -1018,168 +878,16 @@ export default function CreateCampaignPage() {
           </div>
         </form>
 
+        {/* Review modal */}
         {isReviewOpen && reviewData && (
-          <div
-            className="fixed inset-0 z-50 flex items-center justify-center bg-black/60 backdrop-blur-sm px-4"
-            role="dialog"
-            aria-modal="true"
-            aria-labelledby="campaign-review-title"
-            onClick={(e) => {
-              if (e.target === e.currentTarget && !isSubmitting) {
-                setIsReviewOpen(false);
-              }
-            }}
-            onKeyDown={(e) => {
-              if (e.key === "Escape" && !isSubmitting) {
-                setIsReviewOpen(false);
-              }
-            }}
-          >
-            <div className="w-full max-w-xl rounded-2xl border border-zinc-200 dark:border-zinc-700 bg-white dark:bg-zinc-900 shadow-2xl overflow-hidden">
-              <div className="px-6 py-5 border-b border-zinc-200 dark:border-zinc-700">
-                <h2
-                  id="campaign-review-title"
-                  className="text-xl font-semibold text-zinc-900 dark:text-zinc-50"
-                >
-                  {t("reviewTitle")}
-                </h2>
-                <p className="text-sm text-zinc-600 dark:text-zinc-400 mt-1">
-                  {t("reviewSubtitle")}
-                </p>
-              </div>
-
-              <dl className="px-6 py-5 space-y-4">
-                <div className="rounded-xl border border-zinc-200 dark:border-zinc-700 bg-zinc-50 dark:bg-zinc-800/60 p-3">
-                  <dt className="text-xs uppercase tracking-wide text-zinc-500 dark:text-zinc-400">
-                    {t("reviewFieldTitle")}
-                  </dt>
-                  <dd className="text-sm font-medium text-zinc-900 dark:text-zinc-100 mt-1">
-                    {reviewData.title}
-                  </dd>
-                </div>
-
-                <div className="rounded-xl border border-zinc-200 dark:border-zinc-700 bg-zinc-50 dark:bg-zinc-800/60 p-3">
-                  <dt className="text-xs uppercase tracking-wide text-zinc-500 dark:text-zinc-400">
-                    {t("reviewFieldCreatorEmail")}
-                  </dt>
-                  <dd className="text-sm font-medium text-zinc-900 dark:text-zinc-100 mt-1">
-                    {reviewData.creatorEmail || t("reviewCreatorEmailNone")}
-                  </dd>
-                </div>
-
-                <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-                  <div className="rounded-xl border border-zinc-200 dark:border-zinc-700 bg-zinc-50 dark:bg-zinc-800/60 p-3">
-                    <dt className="text-xs uppercase tracking-wide text-zinc-500 dark:text-zinc-400">
-                      {t("reviewFieldFundingGoal")}
-                    </dt>
-                    <dd className="text-sm font-medium text-zinc-900 dark:text-zinc-100 mt-1">
-                      {reviewData.fundingGoalXlm.toLocaleString(undefined, {
-                        maximumFractionDigits: 7,
-                      })}{" "}
-                      XLM
-                    </dd>
-                  </div>
-
-                  <div className="rounded-xl border border-zinc-200 dark:border-zinc-700 bg-zinc-50 dark:bg-zinc-800/60 p-3">
-                    <dt className="text-xs uppercase tracking-wide text-zinc-500 dark:text-zinc-400">
-                      {t("reviewFieldDuration")}
-                    </dt>
-                    <dd className="text-sm font-medium text-zinc-900 dark:text-zinc-100 mt-1">
-                      {t("reviewFieldDurationDays", { count: reviewData.durationDays })}
-                    </dd>
-                  </div>
-                </div>
-
-                <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-                  <div className="rounded-xl border border-zinc-200 dark:border-zinc-700 bg-zinc-50 dark:bg-zinc-800/60 p-3">
-                    <dt className="text-xs uppercase tracking-wide text-zinc-500 dark:text-zinc-400">
-                      {t("reviewFieldCategory")}
-                    </dt>
-                    <dd className="text-sm font-medium text-zinc-900 dark:text-zinc-100 mt-1">
-                      {CATEGORY_LABELS[reviewData.category]}
-                    </dd>
-                  </div>
-
-                  <div className="rounded-xl border border-zinc-200 dark:border-zinc-700 bg-zinc-50 dark:bg-zinc-800/60 p-3">
-                    <dt className="text-xs uppercase tracking-wide text-zinc-500 dark:text-zinc-400">
-                      {t("reviewFieldRevenueShare")}
-                    </dt>
-                    <dd className="text-sm font-medium text-zinc-900 dark:text-zinc-100 mt-1">
-                      {reviewData.hasRevenueSharing
-                        ? `${reviewData.revenueSharePercentage.toFixed(2)}%`
-                        : t("reviewRevenueShareNone")}
-                    </dd>
-                  </div>
-                </div>
-
-                <div className="rounded-xl border border-zinc-200 dark:border-zinc-700 bg-zinc-50 dark:bg-zinc-800/60 p-3">
-                  <dt className="text-xs uppercase tracking-wide text-zinc-500 dark:text-zinc-400">
-                    {t("reviewFieldEndDate")}
-                  </dt>
-                  <dd className="text-sm font-medium text-zinc-900 dark:text-zinc-100 mt-1">
-                    {formatReviewDate(reviewData.estimatedDeadlineTimestamp)}
-                  </dd>
-                </div>
-
-                {reviewData.tags.length > 0 && (
-                  <div className="rounded-xl border border-zinc-200 dark:border-zinc-700 bg-zinc-50 dark:bg-zinc-800/60 p-3">
-                    <dt className="text-xs uppercase tracking-wide text-zinc-500 dark:text-zinc-400">
-                      {t("reviewFieldTags")}
-                    </dt>
-                    <dd className="flex flex-wrap gap-2 mt-1.5">
-                      {reviewData.tags.map((tag) => (
-                        <span
-                          key={tag}
-                          className="px-2 py-0.5 rounded bg-zinc-200 dark:bg-zinc-700 text-zinc-700 dark:text-zinc-300 text-[10px] font-bold border border-zinc-300 dark:border-zinc-600"
-                        >
-                          #{tag}
-                        </span>
-                      ))}
-                    </dd>
-                  </div>
-                )}
-
-                <div className="rounded-xl border border-zinc-200 dark:border-zinc-700 bg-zinc-50 dark:bg-zinc-800/60 p-3">
-                  <dt className="text-xs uppercase tracking-wide text-zinc-500 dark:text-zinc-400">
-                    {t("reviewFieldTimestamp")}
-                  </dt>
-                  <dd className="text-sm font-semibold text-zinc-900 dark:text-zinc-100 mt-1 tabular-nums">
-                    {reviewData.estimatedDeadlineTimestamp}
-                  </dd>
-                </div>
-              </dl>
-
-              <div className="px-6 py-5 border-t border-zinc-200 dark:border-zinc-700 flex flex-col-reverse sm:flex-row sm:items-center sm:justify-between gap-3">
-                <button
-                  type="button"
-                  onClick={() => setIsReviewOpen(false)}
-                  disabled={isSubmitting}
-                  className="px-4 py-2.5 rounded-xl border border-zinc-200 dark:border-zinc-600 bg-white dark:bg-zinc-800 text-zinc-700 dark:text-zinc-200 text-sm font-medium hover:bg-zinc-50 dark:hover:bg-zinc-700 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-                >
-                  {t("editDetails")}
-                </button>
-                <button
-                  type="button"
-                  onClick={handleConfirmAndSign}
-                  disabled={isSubmitting || !isWalletConnected}
-                  className="inline-flex items-center justify-center gap-2 px-5 py-2.5 rounded-xl bg-blue-600 text-white text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
-                >
-                  {isSubmitting && (
-                    <span className="inline-block motion-safe:animate-spin rounded-full h-4 w-4 border-2 border-white border-t-transparent" />
-                  )}
-                  {isSubmitting
-                    ? txPhase === "building"
-                      ? t("submitting")
-                      : txPhase === "signing"
-                        ? "Signing…"
-                        : txPhase === "confirming"
-                          ? "Confirming…"
-                          : t("submitting")
-                    : t("confirmAndSign")}
-                </button>
-              </div>
-            </div>
-          </div>
+          <CampaignReviewModal
+            reviewData={reviewData}
+            isSubmitting={isSubmitting}
+            txPhase={txPhase}
+            onClose={() => setIsReviewOpen(false)}
+            onConfirm={handleConfirmAndSign}
+            formatReviewDate={formatReviewDate}
+          />
         )}
       </main>
     </div>
